requirements.txt
Flask==3.0.0
waitress==2.1.2
Pillow==10.1.0
ultralytics==8.0.213